+++ include/vnx/Hash64.hpp
@@ -19,10 +19,10 @@
 
 #include <vnx/package.hxx>
 #include <vnx/Hash64.h>
-#include <vnx/Hash128.h>
+#include <vnx/Hash128.hpp>
 #include <vnx/Util.h>
-#include <vnx/Input.h>
-#include <vnx/Output.h>
+#include <vnx/Input.hpp>
+#include <vnx/Output.hpp>
 #include <vnx/Visitor.h>
 #include <vnx/CRC64.h>